+++ av1/decoder/decodeframe.h
@@ -20,11 +20,11 @@ extern "C" {
 #if CONFIG_THROUGHPUT_ANALYSIS
 extern int64_t tot_ctx_syms;
 extern int64_t tot_bypass_syms;
-extern int max_ctx_syms;
-extern int max_bypass_syms;
-extern int max_bits;
+extern int64_t max_ctx_syms;
+extern int64_t max_bypass_syms;
+extern int64_t max_bits;
 extern int64_t tot_bits;
-extern int tot_frames;
+extern int64_t tot_frames;
 #endif  // CONFIG_THROUGHPUT_ANALYSIS
 
 struct AV1Decoder;
